+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- **feat**: Added `ctx.resHeaders` readonly getter, which allows you to introspect the current set of outbound headers
+- **feat**: Added `ctx.delHeaders` allowing you to remove multiple outbound headers at once by passing an array of strings
+```typescript
+ctx.setHeader('x-foo', 'bar');
+ctx.setHeader('x-bar', 'baz');
+ctx.setHeader('x-baz', 'qux');
+
+ctx.delHeaders(['x-foo', 'x-baz']);
+
+ctx.resHeaders; /* {'x-bar': 'baz'} */
+```
+
 ### Improved
 - **qol**: `ctx.render` now merges the passed options on top of the defaults
 - **qol**: Inbound and outbound headers will now be treated as case insensitive and be lowercased (as per [RFC 7230](https://datatracker.ietf.org/doc/html/rfc7230#section-3.2) and [RFC 9110](https://datatracker.ietf.org/doc/html/rfc9110))